refactor projectkeydetails
