+++ src/sentry/static/sentry/app/views/settings/project/projectKeys/list/index.tsx
